fix: a race condition which caused hang on stopping
If you supply 'stop' command to stdin right after you started the server,
you may notice the server hangs occasionally. That turns out to be
we requested the server to stop when server is 'setting up' itself,
and it does not think itself 'running' at that time. Also, the run()
function runs in an independent thread, which makes a race condition
to occur. The fix just blocks the main thread until the server goes
into 'running' state, so that it could process 'stop' command correctly.